Add working jest/browser/storybook mocking
