+++ package.json
@@ -10,8 +10,8 @@
     "prepare": "husky install",
     "pre-commit-lint": "yarn compile && yarn lint-staged",
     "test": "jest --config ./jest.config.js --maxWorkers=50%",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "storybook": "start-storybook -s ./packages/host/public  -p 6006",
+    "build-storybook": "build-storybook -s ./packages/host/public "
   },
   "workspaces": {
     "packages": [
@@ -45,6 +45,7 @@
     "jest": "^27.0.6",
     "lerna": "^3.22.1",
     "msw": "^0.35.0",
+    "msw-storybook-addon": "^1.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
