+++ packages/host/src/bootstrap.tsx
@@ -4,9 +4,9 @@ import App from './App';
 
 if (process.env.NODE_ENV === 'development') {
   // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const { worker } = require('@openmsupply-client/mocks');
+  const { setupMockWorker } = require('@openmsupply-client/mocks/src/client');
+  const worker = setupMockWorker();
   worker.start();
 }
 
 ReactDOM.render(<App />, document.getElementById('root'));
-2;
