+++ packages/mocks/src/client.ts
@@ -1,0 +1,4 @@
+import { setupWorker, SetupWorkerApi } from 'msw';
+import { handlers } from './handlers';
+
+export const setupMockWorker = (): SetupWorkerApi => setupWorker(...handlers);
