+++ packages/mocks/src/server.ts
@@ -1,0 +1,4 @@
+import { setupServer, SetupServerApi } from 'msw/node';
+import { handlers } from './handlers';
+
+export const setupMockServer = (): SetupServerApi => setupServer(...handlers);
