+++ packages/transactions/src/OutboundShipment/DetailView/index.ts
@@ -1,0 +1,1 @@
+export { OutboundShipmentDetailView } from './DetailView';
