+++ packages/transactions/src/OutboundShipment/ListView/index.ts
@@ -1,0 +1,1 @@
+export { OutboundShipmentListView } from './ListView';
